refactor(Tree): change indentation width to 20
