+++ src/components/Tree/helpers/constants.ts
@@ -1,0 +1,4 @@
+/* (c) Copyright Frontify Ltd., all rights reserved. */
+
+export const ROOT_ID = '__ROOT__';
+export const INDENTATION_WIDTH = 20;
